refresh logic of deleting car

+++ src/main/java/ru/ryazancev/parkingreservationsystem/services/impl/CarServiceImpl.java
@@ -69,6 +69,9 @@ public class CarServiceImpl implements CarService {
         if (carRepository.findById(carId).isEmpty()) {
             throw new IllegalStateException("Car does not exists");
         }
+        if (carRepository.findReservationByCarId(carId).isPresent()) {
+            throw new IllegalStateException("Can not delete car, because car have reservations");
+        }
         carRepository.delete(carId);
     }
 }
